+++ src/routes/championships/create-championship.ts
@@ -2,7 +2,7 @@ import { prisma } from '@/lib/prisma'
 import { FastifyInstance } from 'fastify'
 import { ZodTypeProvider } from 'fastify-type-provider-zod'
 import z from 'zod'
-import { BadRequestError } from './_errors/BadRequest'
+import { BadRequestError } from '../_errors/BadRequest'
 
 export async function createChampionship(app: FastifyInstance) {
     app.withTypeProvider<ZodTypeProvider>().post(
