+++ src/routes/championships/get-championships-list.ts
@@ -2,7 +2,7 @@ import { prisma } from '@/lib/prisma'
 import { FastifyInstance } from 'fastify'
 import { ZodTypeProvider } from 'fastify-type-provider-zod'
 import z from 'zod'
-import { NotFoundError } from './_errors/NotFound'
+import { NotFoundError } from '../_errors/NotFound'
 
 export async function getChampionshipsList(app: FastifyInstance) {
     app.withTypeProvider<ZodTypeProvider>().get(
