+++ src/routes/users/create-user.ts
@@ -1,9 +1,9 @@
 import { FastifyInstance } from 'fastify'
 import { ZodTypeProvider } from 'fastify-type-provider-zod'
 import z from 'zod'
-import { prisma } from '../lib/prisma'
-import { generateHash } from '../utils/hash'
-import { BadRequestError } from './_errors/BadRequest'
+import { prisma } from '../../lib/prisma'
+import { generateHash } from '../../utils/hash'
+import { BadRequestError } from '../_errors/BadRequest'
 
 export async function createUser(app: FastifyInstance) {
     app.withTypeProvider<ZodTypeProvider>().post(
